Force nuget install at create time not first build
This is due to the requirement to get the deps bundle working the
deps.tsx file needs to be compiled. Since VS, through csproj, is taking
care of the building of TSX files, we need to run msbuild AFTER jspm
install and typings install is complete. Then once the first build
succeeds, and deps.js is created, can we generate/bundle deps.lib.js.